feat: add compositor and layer manager classes
